+++ convertkit/CHANGELOG.md
@@ -1,3 +1,9 @@
+### 3.2.4 2026-04-13
+* Added: Settings: Revoke Access and Refresh Tokens, and remove from settings, when `Disconnect` button clicked
+* Added: Uninstall: Revoke Access and Refresh Tokens, and remove from settings, when the Plugin is deleted (Plugins > Kit > Delete)
+* Added: Remove v3 API Secret if detected in settings
+* Updated: WordPress Libraries to 2.1.5
+
 ### 3.2.3 2026-03-31
 * Fix: Downgrade WordPress Libraries to 2.1.3, resolving issues in 3.2.2 and 3.2.2.1